// seanpeak-dev/src/components/Main/PostList.tsx
import React, { FunctionComponent, useMemo } from 'react'
import PostItem from './PostItem'
import { PostListItemType } from 'types/PostItem.types'
import useInfiniteScroll, {
  useInfiniteScrollType,
} from 'hooks/useInfiniteScroll'

type PostListProps = {
  selectedCategory: string
  posts: PostListItemType[]
}
export type PostType = {
  node: {
    id: string
    frontmatter: {
      title: string
      summary: string
      date: string
      categories: string[]
    }
  }
}

/* const POST_ITEM_DATA = {
  title: '가장 빠르게 지식과 스킬을 쌓는 방법',
  date: '2021.10.19.',
  categories: ['Web', 'Frontend', 'Testing'],
  summary: '인풋이 아닌, 아웃풋 중심으로 노력하기',
  link: '<https://www.google.co.kr/>',
} */

const PostList: FunctionComponent<PostListProps> = function ({
  selectedCategory,
  posts,
}) {
  const { containerRef, postList }: useInfiniteScrollType = useInfiniteScroll(
    selectedCategory,
    posts,
  )

  /*   const postListData = useMemo(
    () =>
      posts.filter(
        ({
          node: {
            frontmatter: { categories },
          },
        }: PostListItemType) =>
          selectedCategory !== 'All'
            ? categories.includes(selectedCategory)
            : true,
      ),
    [selectedCategory],
  ) */

  return (
    <main ref={containerRef}>
      {postList.map(({ node: { id, frontmatter } }: PostListItemType) => (
        <PostItem {...frontmatter} link="https://www.google.co.kr/" key={id} />
      ))}
    </main>
  )
}

export default PostList
